pyproject.toml
[tool.poetry]
name = "scholarr"
version = "0.1.0"
description = "Backend for Scholarr"
authors = ["Sambeg Shrestha <sthasam2@gmail.com>"]
license = "GNU v.3."

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
python-dotenv = "^0.19.2"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.0.0"
pandas = "^1.4.0"
scikit-learn = "^1.0.2"
Pillow = "^9.0.1"
textblob = "^0.17.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
